[amiwm] get the top level gadget window thingy MVP built
gadget_window represents /a/ application window.  Ideally we can have
more than one of them, and they'll have their own glists with gadgets
in there.

This doesn't yet put in the window background, some other useful
window manager attributes and whatnot, I'll do that when I'm not
so tired.  But it's enough to prove that I can cascade events through
a top level glist with the window in it, and then it passes through
the window glist into its gadgets.

I mean it's not that great, but it works, and thus I won't complain
too loudly.  I do have a bunch more work to do here before I call
this /useful/, but hey, progress.   I've been wanting to do this
since I was 19.