[amiwm] plumb some basic test event handling up through the gadget tree.
This is just a test WIP to see if I get events up to requestchoice.c,
which I do.  I'm .. not happy with how inefficient it'll get to keep
recursing each gadget; instead I may want a single event queue with some
ID in a hash table that's quick to lookup.

Anyway this is just a WIP hack to show it works...