[amiwm] migrate Battery module to modules/Battery
In preparation for writing other modules, experiment with migrating
the battery module (which is currently freebsd specific) to modules/Battery/ .

This also ties it into the general build, which is fine for me on FreeBSD
but will fail on non-FreeBSD platforms.  I'll need to go and figure out
some Makefile magic for this next.

In any case, modules can now be put in modules/, so we can start cleaning
up what's going on in the main source dir.