[avcontrol] tinker with a module to create an AV control on freebsd
this is a quick experiment to try and add support for handling the
x86 event key presses for the volume/brightness buttons.

It's very FreeBSD specific - I'm calling the FreeBSD specific commands.
I likely could write code to do this rather than calling system().

.. and I also wonder if I need all of the parser stuff; I could
simplify this greatly if I hard-code the AV events..